refactor(defaults): moved options to defaults module
