+++ src/factory/SentenceFactory.spec.ts
@@ -1,4 +1,4 @@
-import defaults from './defaults';
+import defaults from '../defaults';
 import { SentenceFactory } from './SentenceFactory';
 
 describe('SentenceFactory.js', () => {
